+++ db/statistics.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/protofire/ethpar-beaconchain-explorer/cache"
-	"github.com/protofire/ethpar-beaconchain-explorer/metrics"
+	"github.com/protofire/ethpar-beaconchain-explorer/internal/metrics"
 	"github.com/protofire/ethpar-beaconchain-explorer/price"
 	"github.com/protofire/ethpar-beaconchain-explorer/rpc/consensus"
 	"github.com/protofire/ethpar-beaconchain-explorer/internal/logger"
